#64 CommentContainerコンポーネントの作成
